Create uploader
This is the latest version of this shell script. All variables moved to the top of the file for ease of reuse. I duplicate this file for each of the youtube accounts that I upload to, and then change the variables to suit. This includes the option to set a different log file for each copy of this script that you setup.

I'd like to add variables to allow the script to set the location, license and publish date - but that will need to wait for a future version.