Added default video title variable
This lets you set the default video title - it is used in addition to the filename. 
This is useful in batch upload situations where you might use the name of the event in a number of videos - but still want to keep the filename to allow you to differentiate them. 
Can be left blank if you just want the file name in the resulting youtube video.